refactor(Naming) reducer => index
