+++ src/container/RegionsContainer.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from '../slice/slice';
+} from '../slice';
 
 import { get } from '../utils';
